Rename references
