+++ example/shared/src/commonMain/kotlin/viewmodel/link/LinksViewModel.kt
@@ -29,7 +29,7 @@ import com.splendo.kaluga.architecture.navigation.Navigator
 import com.splendo.kaluga.architecture.navigation.toBundle
 import com.splendo.kaluga.architecture.observable.observableOf
 import com.splendo.kaluga.architecture.viewmodel.NavigatingViewModel
-import com.splendo.kaluga.links.repository.LinksRepoBuilder
+import com.splendo.kaluga.links.LinksBuilder
 import com.splendo.kaluga.resources.localized
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -54,7 +54,7 @@ sealed class BrowserSpecRow : NavigationBundleSpecRow<String>(NavigationBundleSp
 }
 
 class LinksViewModel(
-    linkRepoBuilder: LinksRepoBuilder,
+    linkRepoBuilder: LinksBuilder,
     val builder: AlertPresenter.Builder,
     navigator: Navigator<BrowserNavigationActions<BrowserSpecRow>>
     ) : NavigatingViewModel<BrowserNavigationActions<BrowserSpecRow>>(navigator) {
